+++ mi/ext/commands/bot.py
@@ -74,16 +74,14 @@ class BotBase(GroupMixin):
             return
 
         ctx = await self.get_context(message)
-        return (
+        if ctx.message.content:
             await self.invoke(ctx, *ctx.message.content.split(' '))
-            if ctx.message.content
-            else await self.invoke(ctx)
-        )
+        else:
+            await self.invoke(ctx)
+        await self.dispatch('message', message)
 
-    async def _on_message(self, message):
-        status = await self.process_commands(message)
-        if status is False:
-            await self.dispatch('message', message)
+    async def on_message(self, message):
+        await self.process_commands(message)
 
     def event(self, name=None):
         def decorator(func):
@@ -138,6 +136,8 @@ class BotBase(GroupMixin):
             foo = importlib.import_module(event.__module__)
             coro = getattr(foo, ev)
             await self.schedule_event(coro, event, *args, **kwargs)
+        if ev in dir(self):
+            await self.schedule_event(getattr(self, ev), ev, *args, **kwargs)
         return ev in dir(self)
 
     async def dispatch(self, event_name, *args, **kwargs):
@@ -150,11 +150,6 @@ class BotBase(GroupMixin):
                 foo = importlib.import_module(event.__module__)
                 coro = getattr(foo, ev)
             await self.schedule_event(coro, event, *args, **kwargs)
-        try:
-            coro = getattr(self, ev)
-            await self.schedule_event(coro, ev, *args, **kwargs)
-        except AttributeError:
-            pass
 
     def add_cog(self, cog, override: bool = False) -> None:
         cog_name = cog.__cog_name__