+++ mi/chat.py
@@ -1,13 +1,15 @@
-from typing import List
+from __future__ import annotations
+from typing import TYPE_CHECKING, List
 
 from .abc.chat import AbstractChat, AbstractChatContent
-from .conn import Controller
 from .types.chat import Chat as ChatPayload
-from .user import Author
+from .user import User
 from .utils import api, remove_dict_empty, upper_to_lower
 
 __all__ = ['Chat', 'ChatContent']
 
+if TYPE_CHECKING:
+    from mi.state import ConnectionState
 
 class Chat(AbstractChat):
     """
@@ -67,18 +69,19 @@ class ChatContent(AbstractChatContent):
     チャットオブジェクト
     """
 
-    def __init__(self, data: ChatPayload):
+    def __init__(self, data: ChatPayload, state: ConnectionState):
         self.id: str = data["id"]
         self.created_at: str = data["created_at"]
-        self.content: str = data["content"]
+        self.content: str = data["text"]
         self.user_id: str = data["user_id"]
-        self.author: Author = Author(data["author"])
+        self.author: User = User(data["user"])
         self.recipient_id: str = data["recipient_id"]
         self.recipient: str = data["recipient"]
         self.group_id: str = data["group_id"]
         self.file_id: str = data["file_id"]
         self.is_read: bool = data["is_read"]
         self.reads: List = data["reads"]
+        self._state = state
 
     async def delete(self):
         """
@@ -89,5 +92,5 @@ class ChatContent(AbstractChatContent):
         bool:
             成功したか否か
         """
-        res = await Controller.delete_chat(self.id)
+        res = await self._state._delete_chat(self.id)
         return res.status_code == 204