+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- `Client` に `post_note` メソッドを追加しました。今後のノート投稿はこちらを使用してください
 - `rich` を使ったデバッグ機能を追加しました
 - `file_upload` 関数を追加しました
 - `run` メソッドに `debug` 引数を追加しました
@@ -17,14 +18,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - TypedDictを追加しました
 - チャットをする際に用いる `Chat` `ChatContent` クラスを追加しました
 - `on_chat` イベントを追加しました
-- `NoteContent` クラスを追加しました
-    - ノートの受信イベントでは基本このクラスが使用されます。
+- ~~`NoteContent` クラスを追加しました~~
+    ~~- ノートの受信イベントでは基本このクラスが使用されます。~~
 
 ### Changed
 
 - api周りでのエラー出力が分かりやすくなりました
-- 今までの `Note` クラスに当たるものを `NoteContent` に変更し別途 `Note` クラスを作成させるようにしました
-    - これは不要なデータなどを生成しないようにすることが目的です。
+- ~~今までの `Note` クラスに当たるものを `NoteContent` に変更し別途 `Note` クラスを作成させるようにしました~~
+    - ~~これは不要なデータなどを生成しないようにすることが目的です。~~
 - 部分的にPydanticを廃止
     - 素直に使わないほうが部分的に楽だから
 - `Drive` クラスの `upload` メソッドで例外 `InvalidParameters` を発生させるようにしました
@@ -44,6 +45,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - __on_error が動かない不具合を修正しました
 - 一部の誤った typing hintを修正
 - バグの原因になるコードを修正
+
+### Removed
+
+- `NoteContent` クラスを削除しました
+- `UserAction` クラスを削除しました
+- `UserProfile` クラスを削除しました
+    - これらは `User` に統合され、詳細な情報は `User` クラスの detailsオブジェクト(`UserDetails`) から取得可能です
 
 ## [v0.2.5] 2021-10-07